requirements.txt
flask==3.0.0
anthropic==0.34.2
python-dotenv==1.0.0
requests==2.31.0
gunicorn==21.2.0
flask-cors==4.0.0
markdown==3.5.1
bleach==6.1.0